Try to detect error string in body
The AVR-1912 returns 200 OK on POST to /goform/AppCommand.xml, but with
a error in the body.

Before this patch, that error was missed and a empty mapping for
renamed_sources and deleted_sources was returned, togeather with a
status of True, so _get_renamed_deleted_sources was never run.